+++ lib/output/nsq.go
@@ -5,6 +5,7 @@ import (
 	"github.com/Jeffail/benthos/v3/lib/metrics"
 	"github.com/Jeffail/benthos/v3/lib/output/writer"
 	"github.com/Jeffail/benthos/v3/lib/types"
+	"github.com/Jeffail/benthos/v3/lib/util/tls"
 	"github.com/Jeffail/benthos/v3/lib/x/docs"
 )
 
@@ -24,6 +25,7 @@ batched messages these interpolations are performed per message part.`,
 			docs.FieldCommon("nsqd_tcp_address", "The address of the target NSQD server."),
 			docs.FieldCommon("topic", "The topic to publish to.").SupportsInterpolation(false),
 			docs.FieldCommon("user_agent", "A user agent string to connect with."),
+			tls.FieldSpec(),
 			docs.FieldCommon("max_in_flight", "The maximum number of messages to have in flight at a given time. Increase this to improve throughput."),
 		},
 	}